# Optimized basecalling speed ~2-3x