@nmiculinic 2018-04-28
For some reason, merging master brach  commit:114bc25 breaks the code...investigating