@nmiculinic 2018-04-28
After investigation I've concluded:
* Chiron 0.4 master -- commit a6d0284363100d68d3183eb56092986cb0b6d691 base calling trunk is broken
    *  With command ```TF_CPP_MIN_LOG_LEVEL=2 python chiron/entry.py call -e fasta --beam 0 --batch_size 25 -i chiron/example_data/ -o /tmp/ch3```
        The following error appears: 
```
Traceback (most recent call last):
  File "chiron/entry.py", line 96, in <module>
    main()
  File "chiron/entry.py", line 88, in main
    args.func(args)
  File "chiron/entry.py", line 25, in evaluation
    chiron_eval.run(args)
  File "/home/lpp/Desktop/Chiron/chiron/chiron_eval.py", line 299, in run
    time_dict = unix_time(evaluation)
  File "/home/lpp/Desktop/Chiron/chiron/utils/unix_time.py", line 21, in unix_time
    function(*args, **kwargs)
  File "/home/lpp/Desktop/Chiron/chiron/chiron_eval.py", line 235, in evaluation
    if FLAGS.mode == 'rna':
AttributeError: 'Namespace' object has no attribute 'mode'
```
   * ok, then I try: 
       * ```TF_CPP_MIN_LOG_LEVEL=2 python chiron/chiron_eval.py -e fasta --beam 0 --batch_size 25 -i chiron/example_data/ -o /tmp/ch3```
      and new errror pops out:
```
Traceback (most recent call last):
  File "chiron/chiron_eval.py", line 343, in <module>
    run(args)
  File "chiron/chiron_eval.py", line 299, in run
    time_dict = unix_time(evaluation)
  File "/home/lpp/Desktop/Chiron/chiron/utils/unix_time.py", line 21, in unix_time
    function(*args, **kwargs)
  File "chiron/chiron_eval.py", line 211, in evaluation
    saver.restore(sess, tf.train.latest_checkpoint(FLAGS.model))
  File "/home/lpp/Desktop/Chiron/.venv/lib/python3.6/site-packages/tensorflow-1.8.0rc1-py3.6-linux-x86_64.egg/tensorflow/python/training/saver.py", line 1796, in restore
    raise ValueError("Can't load save_path when it is None.")
ValueError: Can't load save_path when it is None.
```

 Here I gave up. Also, I've noticed master has deleted tqdm, which is quite useful to see how the progress is going on. This PR speeds up the basecalling by 2-3 times (maybe more, further testing needed)