@haotianteng 2018-04-28
Thanks a lot for the help! I will check it for merging.
We have been trying to do this parallel before, glad you make it. 
I have done some benchmark:

**Speed test for beam search decoder:
beam 0 1 2 3 5 10
mean 1.522 1.737 1.982 2.212 2.581 3.461
std 0.028 0.152 0.157 0.161 0.159 0.160

0.194 s/ beam-width*batch(3000*512)*1 CPU

Solution: Use GPU to call Neural network and CPU to call Beam search.
1080Ti + ~ 8 beam width * CPU number**

A ideal setting would be 1080Ti + 4 CPUs with 30 Beam width

However, beam search decoder in Tensorflow does not support multiple threading:
https://github.com/tensorflow/tensorflow/issues/17136
So I am still waiting TF to enable it, but that's awesome you have achieved this! Really appreciate it.